requirements.txt
openai
pydantic
python-dotenv
rich